+++ libft/ft_strlcpy.c
@@ -6,14 +6,14 @@
 /*   By: ledos-sa <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/06 14:53:21 by ledos-sa          #+#    #+#             */
-/*   Updated: 2022/11/07 23:09:07 by ledos-sa         ###   ########.fr       */
+/*   Updated: 2022/11/10 19:31:35 by ledos-sa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stddef.h>
 
-unsigned	int	ft_strlcpy(char *dst, const char *src, size_t size)
+size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
 	size_t	i;
 	size_t	length;
